Move MockPrimaryNode to the subspace-test-service root module
