+++ domains/test/service/src/core_domain.rs
@@ -33,7 +33,7 @@ use std::marker::PhantomData;
 use std::sync::Arc;
 use subspace_runtime_primitives::opaque::Block as PBlock;
 use subspace_runtime_primitives::Index as Nonce;
-use subspace_test_service::mock::MockPrimaryNode;
+use subspace_test_service::MockPrimaryNode;
 use substrate_frame_rpc_system::AccountNonceApi;
 use substrate_test_client::{
     BlockchainEventsExt, RpcHandlersExt, RpcTransactionError, RpcTransactionOutput,
